breedte issues ivm scheidingX
git-svn-id: https://svn.science.uu.nl/repos/project.fisme.java/WiskOpdr/trunk@23608 58dd2db4-0f11-0410-9df4-da15f00f50bd

+++ src/fi/wiskopdr/WiskOpdr.java
@@ -988,8 +988,8 @@ public class WiskOpdr extends JApplet implements ScormAppletIF, ActionListener, 
 		defaultEditModeLaunchData.put("subKnop", new Boolean(false));
 		defaultEditModeLaunchData.put("hasTitle", new Boolean(false));
 		defaultEditModeLaunchData.put("hasAntwoordVak", new Boolean(false));
-		defaultEditModeLaunchData.put("scheidingX", new Integer(780));
-
+		defaultEditModeLaunchData.put("scheidingX", new Integer(1024));
+		
 		defaultEditModeState = StringCodeObject.encodeObjectToString(defaultEditModeLaunchData);
 
 		h.put("aantalActiviteiten", "1");
